--- a/perun/view/flow/factory.py
+++ b/perun/view/flow/factory.py
@@ -3,10 +3,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Hashable, cast, Protocol
 
-import demandimport
-
-with demandimport.enabled():
-    import holoviews as hv
+import holoviews as hv
 
 from perun.profile import convert
 from perun.utils import view_helpers