--- a/pyproject.toml
+++ b/pyproject.toml
@@ -28,9 +28,6 @@ dependencies = [
     "GitPython>=3.1.30",
     "gitdb>=4.0.10",
 
-    # Lazy imports
-    "demandimport>=0.3.4",
-
     # String / text utilities
     "Faker>=19.3",
     "ruamel.yaml>=0.17",
@@ -140,7 +137,6 @@ module = [
     "scipy.*",
     "sklearn.*",
     "statsmodels.*",
-    "demandimport.*",
     "holoviews.*",
     "bcc.*",
     "binaryornot.*",