@JiriPavela 2023-10-09
Thanks for the PR. We used `demandimport` not only for optional dependencies (for which try/except is sufficient), but also for lazy importing to improve Perun startup time. That being said, lazy imports were not solved systematically which explains the still too long startup time.

I was aware that demandimport is deprecated, however, we have not found a replacement for lazy imports yet. [PEP 690](https://peps.python.org/pep-0690/) seemed like it might solve our issue – until it was rejected. So for now, I think removing `demandimport` is the right move (to enable 3.12 support) and the next step should be finding a suitable alternative for lazy imports.