@HarryMichal 2023-10-16
It's better to have concrete values, so let's see.

Before this patch:

```
; time perun
real    0m3,811s
user    0m3,410s
sys     0m0,812s
```

[importtime-pre.txt](https://github.com/Perfexionists/perun/files/12922398/importtime-pre.txt)

After this patch:

```
; time perun
real    0m3,546s
user    0m3,301s
sys     0m0,804s
```

[importtime-post.txt](https://github.com/Perfexionists/perun/files/12922399/importtime-post.txt)

The star-up timing values need to be taken with a pinch of salt as they are dependent on so much but what they show is that removing `demandimport` does not cause a regression in the start-up time.

The `importtime` outputs are much more interesting, though. The biggest offenders seem to be:

- `angr` - ~1.3 second (`unicorn` takes up ~0.5 second of that mainly due to using `distutils` which pulls in `pkg_resources`).
- `matplotlib` - ~0.4 second
- `scipy` - ~0.4 second
- `pandas` - ~0.2 second
- `numpy` - ~0.1 second

Of these only `angr` seems to have some low-hanging fruit to boost import time.