@JiriPavela 2023-10-24
Really interesting results, I believe that `demandimport` originally helped a bit with the startup time but it's quite possible that since it wasn't properly maintained since python 3.7, it might not be lazily importing in some scenarios anymore.

Let's get rid of `demandimport` for now and solve lazy imports later on.