@HarryMichal 2023-10-24
> I believe that `demandimport` originally helped a bit with the startup time but it's quite possible that since it wasn't properly maintained since python 3.7, it might not be lazily importing in some scenarios anymore.

I think it did that but the point of the lazy import is to do it uniformly. One stray import will regress the import time again. E.g., the only times this PR touches imports of `bokeh` is in `perun/utils/view_helpers.py` and `perun/view/scatter/factory.py` but it is also imported in `perun/utils/bokeh_helpers.py`. The same applies to `pandas` and possibly even the others.